+++ crates/blockgen/src/lib.rs
@@ -1,0 +1,6 @@
+mod block;
+mod block_state;
+mod felt;
+mod transaction;
+
+pub use block_state::BlockState;